# [DNM]executor: support tidb_store_batch_size for index_scan rpc